test: extract MockTranscriptionClient.swift from RecordingTestSupport.swift
Splits the `MockTranscriptionClient` actor test double (~43 lines) out
of RecordingTestSupport.swift into its own file so RecordingTestSupport
holds only the primary `MockAudioRecorder` recording double.

Byte-identical move of lines 114-156. Both files retain their test
imports.

RecordingTestSupport.swift: 156 → 112 lines. Tests: 667.
